fix(core): Check that tmp folder is set before delete
The tmp directory is only created if there are files to be
processed. In the event that there are no files, an error
is thrown and "systemExit" is invoked. One of the tasks
of "systemExit" is to delete the tmp folder, which it
fails to do because the tmp directory string hasn't been
set. Try/catch was added because error was being swallowed.

+++ src/nutra.js
@@ -101,7 +101,9 @@ class Private {
     }
 
     systemExit () {
-        Helper.removeDirectory(this.system.tmpDirectory)
+        if (this.system && typeof this.system.tmpDirectory === 'string') {
+            Helper.removeDirectory(this.system.tmpDirectory)
+        }
     }
 
     runHooks (hookType, pluginType, parameters) {
@@ -294,7 +296,11 @@ class Private {
             console.error(err.stack)
         }
         if (!warning || exit) {
-            this.systemExit()
+            try {
+                this.systemExit()
+            } catch (e) {
+                console.error(err.stack)
+            }
             Exit(1)
         }
     }
